requirements.txt
arcgis>=2.4.0
numpy<2
Pillow
tensorflow-macos
tensorflow-metal
